ppm: only emit integer rgb values

+++ src/main.rs
@@ -50,9 +50,9 @@ impl Image {
         for row in &self.pixels {
             for pixel in row {
                 println!("{} {} {}",
-                         pixel.r * 255.0,
-                         pixel.g * 255.0,
-                         pixel.b * 255.0);
+                         (pixel.r * 255.0) as u8,
+                         (pixel.g * 255.0) as u8,
+                         (pixel.b * 255.0) as u8);
             }
         }
     }
